Move some function from cg_llvm/back/link.rs to cg_ssa/back/link.rs
